chore: update checks
